RequireAuth para um componente e manter o AppRoutes focado nas definições de rotas
